requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
openpyxl==3.1.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
Werkzeug==3.0.1
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
python-dotenv
